state_object.go: separate coreth specifics
